[refactor] File...Repository : 직렬화 파일 경로를 캡슐화
- enum 으로 묶으로써 외부에서 직접 접근을 제한하여 데이터를 보호
